+++ chevah/compat/tests/domain/__init__.py
@@ -27,6 +27,7 @@ TEST_USERS = [
         domain=TEST_DOMAIN,
         pdc=TEST_PDC,
         create_profile=True,
+        shared=True,
         )
     ]
 
@@ -68,4 +69,7 @@ def setup_package():
 
 
 def teardown_package():
+    for user in TEST_USERS:
+        user.shared = False
+
     teardown_access_control(users=TEST_USERS, groups=TEST_GROUPS)