+++ chevah/compat/tests/domain/test_filesystem.py
@@ -3,6 +3,9 @@
 """
 Tests for portable filesystem access.
 """
+import win32security
+
+
 from chevah.compat.testing import (
     manufacture,
     TEST_ACCOUNT_GID,
@@ -12,59 +15,33 @@ from chevah.compat.testing import (
     TestUser,
     )
 
-from chevah.compat.administration import os_administration
-from chevah.compat.testing import FileSystemTestCase
+from chevah.compat.testing import OSAccountFileSystemTestCase
 from chevah.compat.tests.mixin.filesystem import SymbolicLinksMixin
 
 
-class SymbolicLinkTestCase(FileSystemTestCase):
+class SymbolicLinkTestCase(OSAccountFileSystemTestCase):
     """
-    Common test case for symbolic link(s) tests.
+    Symbolic link(s) test case.
+
+    User requires SE_CREATE_SYMBOLIC_LINK privilege on Windows OSes
+    in order to be able to create symbolic links.
+
+    We are using a custom user for which we make sure the right is present
+    for these tests.
     """
-
-    @classmethod
-    def setUpTestUser(cls):
-        """
-        Set-up OS user for symbolic link testing.
-
-        User requires SE_CREATE_SYMBOLIC_LINK privilege on Windows OSes
-        in order to be able to create symbolic links.
-
-        We are using a custom user for which we make sure the right is present
-        for these tests.
-        """
-        if cls.os_family != 'nt':
-            raise AssertionError('Only Windows DC clients supported.')
-
-        import win32security
-        rights = (win32security.SE_CREATE_SYMBOLIC_LINK_NAME,)
-
-        username = manufacture.string()
-        user = TestUser(
-            name=username,
-            password=manufacture.string(),
-            domain=TEST_DOMAIN,
-            pdc=TEST_PDC,
-            home_group=TEST_ACCOUNT_GROUP,
-            home_path=u'/home/%s' % username,
-            posix_uid=3000 + manufacture.number(),
-            posix_gid=TEST_ACCOUNT_GID,
-            create_profile=True,
-            windows_required_rights=rights,
-            )
-
-        os_administration.addUser(user)
-
-        return user
-
-    @classmethod
-    def tearDownClass(cls):
-        """
-        Remove OS user used for testing.
-        """
-        os_administration.deleteUser(cls.os_user)
-
-        super(SymbolicLinkTestCase, cls).tearDownClass()
+    _username = manufacture.string()
+    TEST_USER = TestUser(
+        name=_username,
+        password=manufacture.string(),
+        domain=TEST_DOMAIN,
+        pdc=TEST_PDC,
+        home_group=TEST_ACCOUNT_GROUP,
+        home_path=u'/home/%s' % _username,
+        posix_uid=3000 + manufacture.number(),
+        posix_gid=TEST_ACCOUNT_GID,
+        create_profile=True,
+        windows_required_rights=(win32security.SE_CREATE_SYMBOLIC_LINK_NAME,),
+        )
 
 
 class TestSymbolicLink(SymbolicLinkTestCase, SymbolicLinksMixin):