+++ chevah/compat/tests/elevated/__init__.py
@@ -45,4 +45,7 @@ def setup_package():
 
 
 def teardown_package():
+    for user in TEST_USERS:
+        user.shared = False
+
     teardown_access_control(users=TEST_USERS, groups=TEST_GROUPS)