+++ chevah/compat/tests/normal/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (c) 2010 Adi Roiban.
+# See LICENSE for details.
+"""
+Compat tests.
+"""